Refactor signal handling logic for multiprocess and reloader in ProcessTracker
