+++ uvicorn/workers.py
@@ -2,6 +2,7 @@ import asyncio
 import logging
 
 from gunicorn.workers.base import Worker
+
 from uvicorn.config import Config
 from uvicorn.main import Server
